fix: some security
